==> scripts/03_cumulative_release.ipynb <==
<span STYLE="font-size:150%"> 
    Plot drug release parameters
</span>

Docker image: gnasello/datascience-env:2023-01-27 \
Latest update: 20 March 2023

# Load libraries

library(ggplot2)
library(latex2exp)
library(dplyr)
library(tidyverse)

Clone the [r_utils library](https://github.com/gabnasello/r_utils.git) from GitHub in the same folder of this script. 

You can simply running the following command in a new terminal (open it from JupyterLab):

`git clone https://github.com/gabnasello/r_utils.git`

How to Reuse Functions That You Create In Scripts, [tutorial](https://www.earthdatascience.org/courses/earth-analytics/multispectral-remote-sensing-data/source-function-in-R/)

source("r_utils/ggplot_utils.R")

# Reading data

<span style="color:red">**User Input**</span>

filename <- "data/2023-03-28_GN011_release.csv"

# Import the data and look at the first six rows
df <- read.csv(file = filename)
head(df)

df <- df %>% select(c(group, sample, release, day, Laponite))
head(df)

df_test <- df %>% 
                group_by(group)%>% 
                    spread(sample, release)
df_test

## Split data frame by groups 

From [dplyr tutorial](https://dplyr.tidyverse.org/reference/group_split.html)

df_spread <- df %>% 
                group_by(group)%>% 
                  spread(sample, release)

df_split <- group_split( df_spread )
df_split

# Calculate Cumulative Sum by Group

From [statology.org](https://www.statology.org/r-cumulative-sum-by-group/)

df <- df %>% 
        group_by(group, sample) %>% 
                mutate(cum_sum = cumsum(release))
head(df)

fileoutput <- "2023-03-28_GN011_data.csv"

write_csv(df, file = fileoutput)

## Split data frame by groups 

Drop `amount` column from dataframe

df_cum = subset(df, select = -c(release) )
head(df_cum)

From [dplyr tutorial](https://dplyr.tidyverse.org/reference/group_split.html)

df_spread <- df_cum %>% 
                group_by(group)%>% 
                  spread(sample, cum_sum)

df_split <- group_split( df_spread )
df_split
